Change 'include Stalker' to 'extend' in bin/stalk
Calling 'include Stalker' in the context of main (which is an instance
of Object) also extends Object with the class/instance methods from
Stalker, which basically pollutes all objects in an application.

Here is a gist which demonstrates the behaviour:

https://gist.github.com/2231315

And this ticket which is probably related:

https://github.com/adamwiggins/stalker/pull/14

Calling 'extend Stalker' works on the instance of the current object.
When used in the context of main, all the methods from Stalker are
loaded and the global Object isn't polluted.